+++ CHANGELOG.md
@@ -3945,3 +3945,4 @@
 - [2026-08-26 14:21] refactor(state): migrate state controller to modern notifier syntax
 - [2026-08-26 14:30] feat(i18n): integrate custom telemetry logger in i18n
 - [2026-08-26 14:44] refactor(security): migrate security controller to modern notifier syntax
+- [2026-08-26 15:05] feat(network): integrate custom telemetry logger in network